% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supports.R
\name{find.imppoints}
\alias{find.imppoints}
\title{time series extrema using important point algorithm}
\usage{
find.imppoints(x, tolerance = 0.02, lookback = 20)
}
\arguments{
\item{x}{xts object or vector, representing a time series}

\item{tolerance}{threshold for percentage change for extreme points}

\item{lookback}{Used for volatility dependent adaptive threshold}
}
\value{
important points data object (object of class imppoints)
}
\description{
time series extrema using important point algorithm
}

