% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supports.R
\name{find.imppoints}
\alias{find.imppoints}
\title{time series extrema using important point algorithm}
\usage{
find.imppoints_pt(x, tolerance_pt = 10)
}
\arguments{
\item{x}{xts object or vector, representing a time series}

\item{tolerance_pt}{threshold for N points}

}
\value{
important points data object (object of class imppoints)
}
\description{
time series extrema using important point algorithm
}
\examples{
x <- quantmod::getSymbols("^GSPC", auto.assign = FALSE)
x <- x["2015/"]
imppts <- find.imppoints_pt(x,10)
quantmod::chart_Series(x)
points(as.numeric(imppts$maxima$pos),as.numeric(imppts$maxima$value),bg="green",pch=24,cex=1.25)
points(as.numeric(imppts$minima$pos),as.numeric(imppts$minima$value),bg="red",pch=25,cex=1.25)
}
