% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlines.R
\name{find.lines}
\alias{find.lines}
\title{Find major lines from a set of given extremum points}
\usage{
find.lines(x, flag, r.tol = 0.02, theta.tol = 2)
}
\arguments{
\item{x}{a dataframe, with column names x and y}

\item{flag}{set to 1 for finding handling maximum points and -1 for minimum points}

\item{r.tol}{minimum step for r coordinate quantization}

\item{theta.tol}{minimum step for theta coordinate quantization (degrees)}
}
\description{
Find major lines from a set of given extremum points
}
\seealso{
\code{\link[techchart]{sanitize.lines}}

\code{\link[techchart]{mergelines}}
}

