% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find.pattern}
\alias{find.pattern}
\title{Find the past, current and forming technical patter}
\usage{
find.pattern(x, pattern = pattern.db("HS")[[1]], tolerance = c(0.25, 0.5),
  pip.tolerance = c(1.5, 1.25))
}
\arguments{
\item{x}{xts, dataframe, matrix or vector, representing a time series}
}
\value{
pattern match object (object of class tpattern)
}
\description{
Find the past, current and forming technical patter
}
\seealso{
\code{\link[techchart]{find.tpattern}}
}

