% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlines.R
\name{find.tchannel}
\alias{find.tchannel}
\title{Find most current enveloping lines of a given time series}
\usage{
find.tchannel(x, tolerance = 1.5, n = 3, pscore = (0.05)^2, pfit = 0.85)
}
\arguments{
\item{x}{xts object representing a time series}

\item{tolerance}{tolerance specification for important points, expressed as times the standard deviation}

\item{n}{number of lines to to choose the best ones from}

\item{pscore}{envelope score, defines how best the lines envelopes the time series}

\item{pfit}{fit socre, defines how much the lines deviates from the extreme points it connects}
}
\value{
returns the trend channel object (of type class tchannel)
}
\description{
Find most current enveloping lines of a given time series
}
\examples{
x <- quantmod::getSymbols("^GSPC", auto.assign = FALSE)
x <- x["2016-01-01::2016-09-30"]
quantmod::chart_Series(x)
tchannel <- find.tchannel(x,1.1)
tchannel
quantmod::add_TA(tchannel$xlines$maxlines[[1]],on=1)
quantmod::add_TA(tchannel$xlines$minlines[[1]],on=1)
}
\seealso{
\code{\link[techchart]{find_lines}}
}

