% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find.tpattern}
\alias{find.tpattern}
\title{Find matching patterns in a time series}
\usage{
find.tpattern(x, pattern = pattern.db("HS")[[1]], tolerance = 0.25,
  pip.tolerance = 2, find.all = TRUE, trend.adjusted = FALSE)
}
\arguments{
\item{x}{xts, a time series in xts format}

\item{pattern}{a list returned from a call to pattern.db()}

\item{tolerance}{tolerance used for pattern matching definition (vol multiple)}

\item{pip.tolerance}{tolerance used for searching important points (vol multiple)}

\item{find.all}{boolean, whether to return all matches or the latest one}

\item{trend.adjusted}{boolean, whether to adjust for trends during matching}
}
\value{
a list of pattern matches (object of type tpattern)
}
\description{
Find matching patterns in a time series
}
\seealso{
\code{\link[techchart]{pattern.db}}

\code{\link[techchart]{find.pattern}}
}

