% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find.trend.channel}
\alias{find.trend.channel}
\title{Find the last best-fit enveloping lines of a given time series}
\usage{
find.trend.channel(x, tscore = (0.25)^2)
}
\arguments{
\item{x}{xts object representing a time series}
}
\value{
returns the trend channel object (of type class tchannel)
}
\description{
Find the last best-fit enveloping lines of a given time series
}
\seealso{
\code{\link[techchart]{find.tchannel}}
}

