% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlines.R
\name{find_lines}
\alias{find_lines}
\title{Find enveloping lines of a given time series}
\usage{
find_lines(x, tolerance = 2, nlines = 10, pfit = NA, pscore = NA,
  r.tol = 0.02, theta.tol = 2, s = 2, merge.tol = 2)
}
\arguments{
\item{x}{xts object representing a time series}

\item{tolerance}{tolerance specification for important points}

\item{nlines}{max number of lines to return}

\item{pfit}{fit socre, default value is NA (no filtering done)}

\item{pscore}{envelope score, default value is NA (no filtering done)}

\item{r.tol}{r (polar) coordinate tolerance in unit square}

\item{theta.tol}{theta coordinate tolerance in degrees}

\item{s}{minimum number less than the # of colinear points on the line}

\item{merge.tol}{tolerance for merging lines with similar start or end points}
}
\description{
Find enveloping lines of a given time series
}
\seealso{
\code{\link[techchart]{find.tchannel}}
}

