% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlines.R
\name{mergelines}
\alias{mergelines}
\title{Merge lines as computed from a find.lines call}
\usage{
mergelines(xlines, tolerance = 2)
}
\arguments{
\item{xlines}{a findlines object}

\item{tolerance}{Maximum number nearby points to merge}
}
\description{
Merge lines as computed from a find.lines call
}
\seealso{
\code{\link[techchart]{find.lines}}

\code{\link[techchart]{sanitize.lines}}
}

