% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{pattern.db}
\alias{pattern.db}
\title{Database of pattern definitions}
\usage{
pattern.db(type = "all")
}
\arguments{
\item{type}{string, any of "HS", "IHS", "BTOP", "BBOT" or "all"}
}
\value{
a list if pattern definition
}
\description{
definition of standard technical patterns
}

