% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlines.R
\name{sanitize.lines}
\alias{sanitize.lines}
\title{Apply filters to lines as computed from a find.lines or merge.lines call}
\usage{
sanitize.lines(xlines, ptheta = 0.9, pscore = 0.3, pfit = 0.95,
  nlines = 10)
}
\arguments{
\item{xlines}{a findlines object}

\item{ptheta}{cosine of the angle of near-vertical lines to filter out}

\item{pscore}{maximum percentile of the scores to keep}

\item{pfit}{minimum (unadjusted) R-square value to filter the fit}

\item{nlines}{maximum number of lines to keep}
}
\description{
Apply filters to lines as computed from a find.lines or merge.lines call
}
\seealso{
\code{\link[techchart]{find.lines}}

\code{\link[techchart]{mergelines}}
}

