% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpttrend.R
\name{cpt.trend}
\alias{cpt.trend}
\title{change point analysis using binary segmentation}
\usage{
cpt.trend(x, Q, minseglen, penalty)
}
\arguments{
\item{x}{xts, dataframe, matrix or vector, representing a time series}

\item{Q}{Maximum number of change points required}

\item{minseglen}{Minimum length of a trend segment}

\item{penalty}{Penalty value, increasing it reduces number of segments}
}
\value{
change points data (object of class cpttrends)
}
\description{
change point analysis using binary segmentation
}
\seealso{
\code{\link[changepoint]{cpt.mean}}
}

