% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find.pattern}
\alias{find.pattern}
\title{Find current and/ or forming technical patter}
\usage{
find.pattern(x, pattern = pattern.db("HS")[[1]], tolerance = c(0.25, 1),
  pip.tolerance = c(1.1, 1.5))
}
\arguments{
\item{x}{xts, a time series in xts format}

\item{pattern}{a list returned from a call to pattern.db()}

\item{tolerance}{a list tolerance used for pattern matching definition (vol multiple)}

\item{pip.tolerance}{a list tolerance used for searching important points (vol multiple)}
}
\value{
a list of pattern matches (object of type tpattern)
}
\description{
Find current and/ or forming technical patter
}
\seealso{
\code{\link[techchart]{pattern.db}}

\code{\link[techchart]{find.tpattern}}
}

