% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supports.R
\name{find.pivots}
\alias{find.pivots}
\title{supports and resitance from the chart of a time series}
\usage{
find.pivots(x, type = c("SR", "FIB"), tolerance = 0.02, strength = 3,
  maxline = 10, lookback = 20)
}
\arguments{
\item{x}{xts object, or vector, representing a time series}

\item{type}{Either FIB (Fibonacci) or SR. SR is based on best fit lines of multiple peaks and troughs}

\item{tolerance}{threshold for percentage change for extreme points}

\item{strength}{minimum number of extreme points defining a support}

\item{maxline}{maximum number of support/ resistance lines to return}

\item{lookback}{Used for volatility dependent adaptive threshold}
}
\value{
support/ resistance object (object of class supports)
}
\description{
supports and resitance from the chart of a time series
}

